refactor: include main into RootLayout file
