+++ src/components/main/about/About.tsx
@@ -2,9 +2,5 @@ import React from 'react';
 import Title from '../../header/Title';
 
 export default function About() {
-  return (
-    <main className="main">
-      <Title>{'About'}</Title>
-    </main>
-  );
+  return <Title>{'About'}</Title>;
 }
