+++ src/components/main/forms/FormPage.tsx
@@ -3,13 +3,11 @@ import './FormPage.scss';
 import Title from '../../header/Title';
 import Form from './Form';
 
-export default class FormPage extends React.Component {
-  render() {
-    return (
-      <main className="main">
-        <Title>{'Forms page'}</Title>
-        <Form />
-      </main>
-    );
-  }
+export default function FormPage() {
+  return (
+    <>
+      <Title>{'Forms page'}</Title>
+      <Form />
+    </>
+  );
 }
