+++ src/components/main/home/Home.tsx
@@ -6,10 +6,10 @@ import SearchBar from './UI/SearchBar';
 
 export default function Home() {
   return (
-    <main className="main">
+    <>
       <Title>{'Home page'}</Title>
       <SearchBar type={'text'} placeholder={'Type to search...'} data-testid={'main-search-bar'} />
       <Cards />
-    </main>
+    </>
   );
 }
