+++ src/components/main/page404/Page404.tsx
@@ -3,8 +3,8 @@ import Title from '../../header/Title';
 
 export default function Page404() {
   return (
-    <main className="main">
+    <>
       <Title>{'Page not found :('}</Title>
-    </main>
+    </>
   );
 }
