--- src/data/data.ts
@@ -1,158 +1,0 @@
-import { TMusic } from '../types/props-types';
-
-const music: TMusic = {
-  albums: [
-    {
-      id: 1,
-      artist: 'Cult of Luna',
-      album: 'Salvation',
-      year: 2004,
-      rating: 10,
-      genre: 'Sludge Post Metal',
-      tracks: 8,
-      cover: 'https://www.darkside.ru/band/336/cover/2406.jpg',
-    },
-    {
-      id: 2,
-      artist: 'Cult of Luna',
-      album: 'Somewhere Along the Highway',
-      year: 2006,
-      rating: 8,
-      genre: 'Sludge Post Metal',
-      tracks: 7,
-      cover: 'https://www.darkside.ru/band/336/cover/9532.jpg',
-    },
-    {
-      id: 3,
-      artist: 'Cult of Luna',
-      album: 'Eternal Kingdom',
-      year: 2008,
-      rating: 9,
-      genre: 'Sludge Post Metal',
-      tracks: 10,
-      cover: 'https://www.darkside.ru/band/336/cover/13566.jpg',
-    },
-    {
-      id: 4,
-      artist: 'Cult of Luna',
-      album: 'Fire Was Born (live)',
-      year: 2009,
-      rating: 10,
-      genre: 'Sludge Post Metal',
-      tracks: 10,
-      cover: 'https://www.darkside.ru/band/336/cover/20166.jpg',
-    },
-    {
-      id: 5,
-      artist: 'Cult of Luna',
-      album: 'Vertikal',
-      year: 2013,
-      rating: 9,
-      genre: 'Sludge Post Metal',
-      tracks: 10,
-      cover: 'https://www.darkside.ru/band/336/cover/33731.jpg',
-    },
-    {
-      id: 6,
-      artist: 'Cult of Luna',
-      album: 'Mariner',
-      year: 2016,
-      rating: 8,
-      genre: 'Sludge Post Metal',
-      tracks: 6,
-      cover: 'https://www.darkside.ru/band/336/cover/42754.jpg',
-    },
-    {
-      id: 7,
-      artist: 'Cult of Luna',
-      album: 'Years in a Day (live)',
-      year: 2017,
-      rating: 9,
-      genre: 'Sludge Post Metal',
-      tracks: 27,
-      cover: 'https://www.darkside.ru/band/336/cover/43930.jpg',
-    },
-    {
-      id: 8,
-      artist: 'Cult of Luna',
-      album: 'A Dawn to Fear',
-      year: 2019,
-      rating: 8,
-      genre: 'Sludge Post Metal',
-      tracks: 8,
-      cover: 'https://www.darkside.ru/band/336/cover/47606.jpg',
-    },
-    {
-      id: 9,
-      artist: 'Cult of Luna',
-      album: 'The Long Road North',
-      year: 2022,
-      rating: 7,
-      genre: 'Sludge Post Metal',
-      tracks: 9,
-      cover: 'https://www.darkside.ru/band/336/cover/49902.jpg',
-    },
-    {
-      id: 10,
-      artist: 'Amenra',
-      album: 'Mass II',
-      year: 2005,
-      rating: 7,
-      genre: 'Post Metal',
-      tracks: 5,
-      cover: 'https://www.darkside.ru/band/6034/cover/34473.jpg',
-    },
-    {
-      id: 11,
-      artist: 'Amenra',
-      album: 'Mass III',
-      year: 2006,
-      rating: 6,
-      genre: 'Post Metal',
-      tracks: 6,
-      cover: 'https://www.darkside.ru/band/6034/cover/15253.jpg',
-    },
-    {
-      id: 12,
-      artist: 'Amenra',
-      album: 'Mass IIII',
-      year: 2008,
-      rating: 7,
-      genre: 'Post Metal',
-      tracks: 7,
-      cover: 'https://www.darkside.ru/band/6034/cover/15254.jpg',
-    },
-    {
-      id: 13,
-      artist: 'Amenra',
-      album: 'Live (live)',
-      year: 2012,
-      rating: 8,
-      genre: 'Post Metal',
-      tracks: 8,
-      cover: 'https://www.darkside.ru/band/6034/cover/34475.jpg',
-    },
-    {
-      id: 14,
-      artist: 'Amenra',
-      album: 'Mass V',
-      year: 2012,
-      rating: 9,
-      genre: 'Post Metal',
-      tracks: 4,
-      cover: 'https://www.darkside.ru/band/6034/cover/34476.jpg',
-    },
-    {
-      id: 15,
-      artist: 'Amenra',
-      album: 'Mass VI',
-      year: 2017,
-      rating: 9,
-      genre: 'Post Metal',
-      tracks: 6,
-      cover: 'https://www.darkside.ru/band/6034/cover/44467.jpg',
-    },
-  ],
-};
-
-export default music;
